:bug: Fix plugin details view
