+++ src/actions/index.js
@@ -1,6 +1,27 @@
 import linkHeader from 'parse-link-header'
 import qs from 'qs'
 
+import {pick} from 'lodash'
+
+const VALID_KEYS = [
+  'id',
+  'name',
+  'title',
+  'description',
+  'version',
+  'compatible_version',
+  'score',
+  'watchers_count',
+  'stargazers_count',
+  'auto_updates',
+  'source_url',
+  'thumbnail_url',
+  'download_url',
+  'owner'
+]
+
+const prunePluginData = (plugins) => plugins.map(p => pick(p, VALID_KEYS))
+
 export const LOCATION_CHANGE = '@@router/LOCATION_CHANGE'
 
 export const PLUGINS_SORT_BY = 'plugins/SORT_BY'
@@ -35,7 +56,7 @@ export const PLUGINS_RECEIVED = 'PLUGINS_RECEIVED'
 export function pluginsReceived (payload) {
   return {
     type: PLUGINS_RECEIVED,
-    payload: payload
+    payload: prunePluginData(payload)
   }
 }
 
@@ -45,7 +66,7 @@ export const PLUGINS_FETCH_RECEIVED = 'plugins/FETCH_RECEIVED'
 export function fetchPluginsReceived (payload) {
   return {
     type: PLUGINS_FETCH_RECEIVED,
-    payload: payload
+    payload: prunePluginData(payload)
   }
 }
 
@@ -131,7 +152,7 @@ export const FETCH_CATALOG_RECEIVED = 'catalog/FETCH_RECEIVED'
 export function fetchCatalogReceived ({payload, append}) {
   return {
     type: FETCH_CATALOG_RECEIVED,
-    payload: payload,
+    payload: prunePluginData(payload),
     append
   }
 }
@@ -163,7 +184,7 @@ export function catalogSortBy (sort) {
     q: '',
     sort: sort
   })
-  
+
   return (dispatch, getState, {api}) => {
 
     dispatch(fetchCatalog(queryParams, false))
@@ -242,7 +263,7 @@ export const FETCH_LIBRARY_RECEIVED = 'library/FETCH_RECEIVED'
 export function fetchLibraryReceived (payload) {
   return {
     type: FETCH_LIBRARY_RECEIVED,
-    payload: payload
+    payload: prunePluginData(payload)
   }
 }
 
@@ -304,7 +325,7 @@ export const FETCH_USER_PLUGINS_RECEIVED = 'user/FETCH_PLUGINS_RECEIVED'
 export function fetchUserPluginsReceived (plugins) {
   return {
     type: FETCH_USER_PLUGINS_RECEIVED,
-    payload: plugins
+    payload: prunePluginData(plugins)
   }
 }
 
@@ -323,7 +344,10 @@ export function fetchUserPlugins (endpoint) {
       .then(response => {
         dispatch(fetchUserPluginsReceived(response.data))
       })
-      .catch(error => dispatch(fetchUserPluginsError(error)))
+      .catch(error => {
+        dispatch(fetchUserPluginsError(error))
+
+      })
   }
 }
 
@@ -338,7 +362,6 @@ export function fetchPluginDetails ({userId, pluginId}) {
 
     api.getPlugin({userId, pluginId})
       .then(response => {
-
         dispatch(fetchPluginDetailsReceived(response.data))
         dispatch(fetchPluginReadme(response.data.readme_url))
       })
